+++ Assets/Scripts/Camara_Registros/Registros/SessionInfo.cs
@@ -7,7 +7,7 @@ public class SessionInfo : MonoBehaviour
 {
     public TxtController DBRegister;
 
-    public int _studentID;
+    public float _studentID;
     public int _sessionID = 0;
     public DateTime _sessionStartTime;
     public DateTime _sessionEndTime;
